+++ packages/navigation/app-navigation/src/components/AppSwitcher/types.ts
@@ -1,5 +1,5 @@
-import { TriggerManagerProps } from '../TriggerManager/types';
+import { IconButtonProps } from '../IconButton/types';
 
-export type AppSwitcherProps = Omit<TriggerManagerProps, 'children'> & {
+export type AppSwitcherProps = Omit<IconButtonProps, 'icon' | 'tooltip'> & {
   tooltip: string;
 };