+++ packages/navigation/app-navigation/src/index.ts
@@ -1,13 +1,17 @@
 export { AppNavigation } from './components/AppNavigation';
 export { AppNavigationSkeleton } from './components/AppNavigation/skeleton';
 export { AppSwitcher } from './components/AppSwitcher';
-export { Create, ThemedCreate } from './components/Create';
-export { Help, ThemedHelp } from './components/Help';
+export { Create } from './components/Create';
+export { Help } from './components/Help';
 export { Notifications } from './components/Notifications';
+export { PrimaryButton } from './components/PrimaryButton';
+export { PrimaryButtonProps } from './components/PrimaryButton/types';
+export { PrimaryDropdownButton } from './components/PrimaryDropdownButton';
 export { CustomProductHome, ProductHome } from './components/ProductHome';
 export { Profile } from './components/Profile';
 export { Search } from './components/Search';
 export { Settings } from './components/Settings';
+export { SignIn } from './components/SignIn';
 
 export {
   Colors,
@@ -15,6 +19,5 @@ export {
   GenerateThemeArgs,
   darkTheme,
   lightTheme,
-  withAppNavigationTheme,
   ThemeProvider,
 } from './theme';