+++ packages/navigation/atlassian-switcher/docs/01-intro.tsx
@@ -17,33 +17,35 @@ export default md`
   * Any cross-flow and admin links, if applicable
   * Any custom links from Jira or Confluence, if applicable
 
-  ## Usage
+  ## Integrating switcher
+
+  There are multiple ways to integrate switcher within your app. Depending on your application you might need to combine some of the options below.
+
+  * [Atlassian switcher vanilla](/packages/navigation/atlassian-switcher-vanilla) can be used in applications that don't bundle React.
+  * [Standalone switcher](/packages/navigation/atlassian-switcher/docs/standalone-switcher) allows to render the switcher in any container (e.g. inline dialog) other than the drawer by specifying the appearance property.
+  * [Custom themes](/packages/navigation/atlassian-switcher/docs/theming-guide) allow to change the colours in the switcher component.
+
+  ## Basic example
+
+  This is a basic example of the switcher being rendered in a drawer.
 
   ${code`import AtlassianSwitcher  from '@atlaskit/atlassian-switcher';`}
 
   ${(
     <Example
-      Component={require('../examples/00-base-example').default}
-      title="Basic"
-      source={require('!!raw-loader!../examples/00-base-example')}
-    />
-  )}
-
-  ## Theming
-
-  ${(
-    <Example
       packageName="@atlaskit/atlassian-switcher"
-      Component={require('../examples/01-standalone-themed-green').default}
-      title="Standalone switcher with green theme"
-      source={require('!!raw-loader!../examples/01-standalone-themed-green')}
+      Component={
+        require('../examples/82-uc-generic-switcher-with-xflow').default
+      }
+      title="Basic switcher example"
+      source={require('!!raw-loader!../examples/82-uc-generic-switcher-with-xflow')}
     />
   )}
 
   ${(
     <Props
       heading="Props"
-      props={require('!!extract-react-types-loader!../src/components/switcher')}
+      props={require('!!extract-react-types-loader!../src/components/atlassian-switcher')}
     />
   )}
 `;