+++ packages/navigation/atlassian-switcher/src/types.ts
@@ -1,5 +1,6 @@
 import { UIAnalyticsEvent } from '@atlaskit/analytics-next';
 import { WithTheme } from './theme/types';
+import { AvailableProductsDataProvider } from './providers/products-data-provider';
 
 export interface TriggerXFlowCallback {
   (
@@ -72,9 +73,20 @@ export enum ProductTopItemVariation {
 }
 
 export type FeatureFlagProps = {
-  [key in Exclude<Feature, typeof Feature.xflow>]: boolean
-} & {
-  [MultiVariateFeature.productTopItemVariation]: ProductTopItemVariation;
+  // Show user centric avaialble products as opposed to site centric product list.
+  enableUserCentricProducts?: boolean;
+  // Custom links are enabled by default for Jira and Confluence, this feature flag allows to hide them. Custom links are not supported by the switcher in any other products.
+  disableCustomLinks?: boolean;
+  // Hide recent containers. Recent containers are enabled by default.
+  disableRecentContainers?: boolean;
+  // Remove section headers - useful if something else is providing them. i.e: trello inline dialog.
+  disableHeadings?: boolean;
+  // Enable discover more.
+  isDiscoverMoreForEveryoneEnabled?: boolean;
+  // Enable Embedded Marketplace within the product.
+  isEmceeLinkEnabled?: boolean;
+  // Defines which site is displayed as the top item for users with multiple sites.
+  [MultiVariateFeature.productTopItemVariation]?: ProductTopItemVariation;
 };
 
 export type FeatureMap = { [key in Feature]: boolean } & {
@@ -115,6 +127,7 @@ export enum WorklensProductType {
   OPSGENIE = 'OPSGENIE',
   BITBUCKET = 'BITBUCKET',
   STATUSPAGE = 'STATUSPAGE',
+  TRELLO = 'TRELLO',
 }
 
 export type AvailableProduct =
@@ -130,7 +143,11 @@ export type AvailableProduct =
 
 interface AvailableProductWithUrl {
   activityCount: number;
-  productType: WorklensProductType.BITBUCKET | WorklensProductType.OPSGENIE;
+  productType:
+    | WorklensProductType.BITBUCKET
+    | WorklensProductType.OPSGENIE
+    | WorklensProductType.STATUSPAGE // assuming that the URL is provided by TCS (same as Opsgenie)
+    | WorklensProductType.TRELLO;
   url: string;
 }
 
@@ -173,9 +190,16 @@ export interface SwitcherChildItem {
 }
 
 export type AtlassianSwitcherProps = WithTheme & {
+  // Product name used for analytics events
   product: string;
+  // Optional cloudID, should be provided for tenanted applications.
   cloudId?: string;
+  // Optional callback to be exectuted after an XFlow event is triggered.
   triggerXFlow?: TriggerXFlowCallback;
+  // Optional callback to be exectuted after a user clicks on discover more.
   onDiscoverMoreClicked?: DiscoverMoreCallback;
+  // A map of feature flags used by the XFlow recommendations engine.
   recommendationsFeatureFlags?: RecommendationsFeatureFlags;
-} & Partial<FeatureFlagProps>;
+  // Optional custom provider for available products
+  availableProductsDataProvider?: AvailableProductsDataProvider;
+} & FeatureFlagProps;