+++ packages/core/right-side-panel/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atlaskit/right-side-panel
 
+## 0.2.0
+
+### Minor Changes
+
+- [minor][617591dd61](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/617591dd61):
+
+  Added props onOpenAnimationFinished and onCloseAnimationFinished to execute function once the open/close animation finished
+
 ## 0.1.8
 
 ### Patch Changes